django/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.1
gunicorn==23.0.0
psycopg[binary]==3.2.9
sqlparse==0.5.3
uvicorn==0.35.0
